Sharity/sharity/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
Pillow==8.4.0
requests>=2.27.1
python-dotenv==0.19.2
python-decouple>=3.5,<4.0